Extract message deletion into separate class
